# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Corporate Logistics Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vehicles:
    get:
      operationId: listVehicles
      tags: [vehicles]
      summary: List all vehicles
      responses:
        "200":
          description: List of vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
    post:
      operationId: createVehicle
      tags: [vehicles]
      summary: Create a vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
  /vehicles/{id}:
    get:
      operationId: getVehicle
      tags: [vehicles]
      summary: Get a vehicle by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "404":
          description: Not found
    patch:
      operationId: updateVehicle
      tags: [vehicles]
      summary: Update a vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "404":
          description: Not found
    delete:
      operationId: deleteVehicle
      tags: [vehicles]
      summary: Delete a vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # --- PERSONNEL ---
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List all sales
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Create a sale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      summary: Get a sale by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
        "404":
          description: Not found
    patch:
      operationId: updateSale
      tags: [sales]
      summary: Update a sale
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
        "404":
          description: Not found
    delete:
      operationId: deleteSale
      tags: [sales]
      summary: Delete a sale
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # --- DISPATCHES ---
...
  /vehicles/recommend:
    post:
      operationId: recommendVehicle
      tags: [vehicles]
      summary: Recommend the best vehicle for a given load
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoadRequirement"
      responses:
        "200":
          description: Recommended vehicles sorted by suitability
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"

  # --- COST ESTIMATION ---
...
  /tolls:
    get:
      operationId: listTolls
      tags: [tolls]
      summary: List all toll routes
      responses:
        "200":
          description: List of toll routes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TollRoute"
    post:
      operationId: createToll
      tags: [tolls]
      summary: Create a toll route
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TollRouteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TollRoute"
  /tolls/{id}:
    patch:
      operationId: updateToll
      tags: [tolls]
      summary: Update a toll route
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TollRouteUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TollRoute"
        "404":
          description: Not found
    delete:
      operationId: deleteToll
      tags: [tolls]
      summary: Delete a toll route
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # --- DASHBOARD ---
...
components:
  schemas:
...
    Vehicle:
      type: object
      required: [id, tipo, modelo, capacidadPeso, capacidadVolumen, tipoCombustible, rendimientoKmLitro, createdAt]
      properties:
        id:
          type: integer
        tipo:
          type: string
          description: "propio | tercero"
        modelo:
          type: string
        capacidadPeso:
          type: number
          description: Max weight capacity in kg
        capacidadVolumen:
          type: number
          description: Max volume capacity in m3
        tipoCombustible:
          type: string
          description: "gasolina | diesel | gas"
        rendimientoKmLitro:
          type: number
          description: Fuel efficiency in km/liter
        placa:
          type: ["string", "null"]
        createdAt:
          type: string
    VehicleInput:
      type: object
      required: [tipo, modelo, capacidadPeso, capacidadVolumen, tipoCombustible, rendimientoKmLitro]
      properties:
        tipo:
          type: string
        modelo:
          type: string
        capacidadPeso:
          type: number
        capacidadVolumen:
          type: number
        tipoCombustible:
          type: string
        rendimientoKmLitro:
          type: number
        placa:
          type: string
    VehicleUpdate:
      type: object
      properties:
        tipo:
          type: string
        modelo:
          type: string
        capacidadPeso:
          type: number
        capacidadVolumen:
          type: number
        tipoCombustible:
          type: string
        rendimientoKmLitro:
          type: number
        placa:
          type: string
    LoadRequirement:
      type: object
      required: [pesoTotal, volumenTotal]
      properties:
        pesoTotal:
          type: number
        volumenTotal:
          type: number

    # --- PERSONNEL ---
...
    Sale:
      type: object
      required: [id, cliente, volumenTotal, pesoTotal, destino, estado, createdAt]
      properties:
        id:
          type: integer
        cliente:
          type: string
        volumenTotal:
          type: number
        pesoTotal:
          type: number
        destino:
          type: string
        estado:
          type: string
          description: "pendiente | despachado | entregado | cancelado"
        notas:
          type: ["string", "null"]
        createdAt:
          type: string
    SaleInput:
      type: object
      required: [cliente, volumenTotal, pesoTotal, destino]
      properties:
        cliente:
          type: string
        volumenTotal:
          type: number
        pesoTotal:
          type: number
        destino:
          type: string
        estado:
          type: string
        notas:
          type: string
    SaleUpdate:
      type: object
      properties:
        cliente:
          type: string
        volumenTotal:
          type: number
        pesoTotal:
          type: number
        destino:
          type: string
        estado:
          type: string
        notas:
          type: string

    # --- DISPATCHES ---
...
    TollRoute:
      type: object
      required: [id, origen, destino, cantidadPeajes, costoTotal]
      properties:
        id:
          type: integer
        origen:
          type: string
        destino:
          type: string
        cantidadPeajes:
          type: integer
        costoTotal:
          type: number
        descripcion:
          type: ["string", "null"]
    TollRouteInput:
      type: object
      required: [origen, destino, cantidadPeajes, costoTotal]
      properties:
        origen:
          type: string
        destino:
          type: string
        cantidadPeajes:
          type: integer
        costoTotal:
          type: number
        descripcion:
          type: string
    TollRouteUpdate:
      type: object
      properties:
        origen:
          type: string
        destino:
          type: string
        cantidadPeajes:
          type: integer
        costoTotal:
          type: number
        descripcion:
          type: string

    # --- DASHBOARD ---
...