# lib/api-zod/src/generated/types/tollRoute.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Corporate Logistics Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface TollRoute {
  id: number;
  origen: string;
  destino: string;
  cantidadPeajes: number;
  costoTotal: number;
  /** @nullable */
  descripcion?: string | null;
}