# lib/api-zod/src/generated/types/vehicleUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Corporate Logistics Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface VehicleUpdate {
  tipo?: string;
  modelo?: string;
  capacidadPeso?: number;
  capacidadVolumen?: number;
  tipoCombustible?: string;
  rendimientoKmLitro?: number;
  placa?: string;
}